feat: Add gitea_gpg_key resource (#106)
Add a new resource to manage a user's GPG keys. Due to a limitation in
Gitea's API, it is only possible to manage GPG keys for the current
user. This is not ideal, but still better than nothing.
If such an API is created, we could extend this resource with an
optional username, while keeping the current behavior as is.

Co-authored-by: techknowlogick <techknowlogick@noreply.gitea.com>
Reviewed-on: https://gitea.com/gitea/terraform-provider-gitea/pulls/106
Reviewed-by: techknowlogick <techknowlogick@noreply.gitea.com>
Co-authored-by: Mathieu Lemay <acidrain1@gmail.com>
Co-committed-by: Mathieu Lemay <acidrain1@gmail.com>